+++ megalinter/reporters/ConsoleLinterReporter.py
@@ -20,6 +20,10 @@ class ConsoleLinterReporter(Reporter):
             self.report_type = "detailed"
         super().__init__(params)
 
+    def manage_activation(self):
+        if config.get("CONSOLE_REPORTER", "true") == "false":
+            self.is_active = False
+
     def produce_report(self):
         linter_version = self.master.get_linter_version()
         # Linter header prints